i refactored dataspec instead of writing docs and im not sorry
