+++ composer/__init__.py
@@ -12,7 +12,6 @@ from composer import utils as utils
 from composer.algorithms import functional as functional
 from composer.core import Algorithm as Algorithm
 from composer.core import Callback as Callback
-from composer.core import DataSpec as DataSpec
 from composer.core import Engine as Engine
 from composer.core import Event as Event
 from composer.core import Logger as Logger
