+++ composer/core/evaluator.py
@@ -2,11 +2,9 @@
 from __future__ import annotations
 
 import copy
-from typing import TYPE_CHECKING, Union
+from typing import TYPE_CHECKING
 
 from torchmetrics import Metric, MetricCollection
-
-from composer.core.data_spec import DataSpec as DataSpec
 
 if TYPE_CHECKING:
     from composer.core.types import DataLoader, Metrics
@@ -17,17 +15,14 @@ class Evaluator:
 
     Attributes:
         label (str): Name of the Evaluator
-        dataloader (Union[DataSpec, DataLoader]): Dataloader/DataSpec for evaluation data
+        dataloader (DataLoader): Dataloader for evaluation data
         metrics (Metrics): Metrics to log. The metrics will be deep-copied to ensure that
             each evaluator updates only its metrics.
     """
 
-    def __init__(self, *, label: str, dataloader: Union[DataSpec, DataLoader], metrics: Metrics):
+    def __init__(self, *, label: str, dataloader: DataLoader, metrics: Metrics):
         self.label = label
-        if isinstance(dataloader, DataSpec):
-            self.dataloader = dataloader
-        else:
-            self.dataloader = DataSpec(dataloader)
+        self.dataloader = dataloader
 
         # Forcing metrics to be a MetricCollection simplifies logging results
         metrics = copy.deepcopy(metrics)
