+++ composer/core/types.py
@@ -15,7 +15,6 @@ from torchmetrics.collections import MetricCollection
 from torchmetrics.metric import Metric
 
 from composer.core.algorithm import Algorithm as Algorithm
-from composer.core.data_spec import DataSpec as DataSpec
 from composer.core.evaluator import Evaluator as Evaluator
 from composer.core.event import Event as Event
 from composer.core.logging import Logger as Logger
